refactor: 💡 PR Ajoute la fiche événement enrichie
